refactor(OnnxToAxonBench): factor out http utils
